requirements.txt
acme==1.14.0
asn1crypto==0.24.0
Babel==2.5.1
certbot==1.14.0
certbot-nginx==1.14.0
cffi==1.11.5
chardet==3.0.4
click==8.0.1
cloud-init==20.3
ConfigArgParse==0.14.0
configobj==5.0.6
cryptography==3.2.1
dataclasses==0.8
dbus-python==1.2.4
decorator==4.2.1
distro==1.4.0
ethtool==0.14
Flask==2.0.1
gpg==1.13.1
gunicorn==20.1.0
idna==2.5
importlib-metadata==4.5.0
iniparse==0.4
itsdangerous==2.0.1
Jinja2==3.0.1
josepy==1.8.0
jsonpatch==1.21
jsonpointer==1.10
jsonschema==2.6.0
libcomps==0.1.16
lxml==4.6.3
MarkupSafe==2.0.1
oauthlib==2.1.0
parsedatetime==2.5
perf==0.1
pexpect==4.3.1
ply==3.9
prettytable==0.7.2
ptyprocess==0.5.2
pyasn1==0.3.7
pycairo==1.16.3
pycparser==2.14
pydbus==0.6.0
Pygments==2.9.0
pygobject==3.28.3
pyinotify==0.9.6
PyJWT==1.6.1
pyOpenSSL==19.0.0
pyparsing==2.1.10
pyRFC3339==1.1
pyserial==3.1.1
PySocks==1.6.8
python-dateutil==2.6.1
python-dmidecode==3.12.2
python-dotenv==0.17.1
python-linux-procfs==0.6.3
pytz==2017.2
pyudev==0.21.0
PyYAML==3.12
requests==2.20.0
requests-toolbelt==0.9.1
rpm==4.14.3
schedutils==0.6
selinux==2.9
sepolicy==1.1
setools==4.3.0
setroubleshoot==1.1
six==1.11.0
slip==0.6.4
slip.dbus==0.6.4
sos==4.1
subscription-manager==1.28.17
syspurpose==1.28.17
systemd-python==234
typing-extensions==3.10.0.0
urllib3==1.24.2
Werkzeug==2.0.1
xq==0.0.4
zipp==3.4.1
zope.component==4.3.0
zope.event==4.2.0
zope.interface==4.6.0
